Improve and fix Correios tests
Now there is a new response that xmerl (used by sweet_xml) cannot handle
well because of latin1-encoded characters, like `Á`. It happens now
because the API started to tell the difference between invalid and not
found CEPs. This requires ai little hack to work well.

To keep backwards compatibility, the invalid ceps will return the
`not_found` error. A different error for invalid CEPs might be
implemented in the future.